requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
yfinance==0.2.65
pandas==2.2.3
openai==1.51.0
httpx==0.27.0
streamlit==1.39.0
python-dotenv==1.0.1
pytest==8.3.3
